(features - make readParamFileAlt return 64-bit array)
Ignore-this: 573328c879fd3afacc4852f678b125d6

darcs-hash:20130507162658-3a4db-3740b36db00e4898aeb262c0acfabd26426ccf90

+++ armspeech/speech/features.py
@@ -48,7 +48,7 @@ def readParamFile(paramFile, paramOrder, decode = None):
 #   than readParamFile!)
 @codeDeps()
 def readParamFileAlt(paramFile, paramOrder):
-    return np.reshape(np.fromfile(paramFile, dtype = np.float32), (-1, paramOrder))
+    return np.reshape(np.fromfile(paramFile, dtype = np.float32), (-1, paramOrder)).astype(np.float64)
 
 @codeDeps()
 def writeParamFile(outSeq, paramFile, paramOrder, encode = None):
